requirements.txt
argh==0.26.1
decorator==4.0.2
ipython==4.0.0
ipython-genutils==0.1.0
nose==1.3.7
nose-watch==0.9.1
path.py==8.1
pathtools==0.1.2
pexpect==3.3
pickleshare==0.5
PyYAML==3.11
requests==2.7.0
simplegeneric==0.8.1
six==1.9.0
spec==1.2.2
traitlets==4.0.0
vcrpy==1.7.3
watchdog==0.8.3
wheel==0.24.0
wrapt==1.10.5
